Try to preserve existing values
If you decode a JSON document onto a struct which already has some values,
it will only write to the fields that are defined in the JSON document. All
other fields are left untouched.

This PR tries to create the same behavior when decoding HCL on a struct
that already contains some (default) values.