gohcl: retain blocks while decoding
Currently, if non-empty struct is passed to DecodeBody function,
decoding process will keep already initialized string fields values or
overwrite them, if they are specified in HCL. This behaviour is useful,
as it allows to have some default values for fields.

However, if the field is a block or multiple blocks (slice), then the
entire block is overwritten, which erases the struct. Because of that,
settings default values in nested structs is not possible.

Consider following HCL and struct example:

foo "foo" {
  nested {
    more = "foo"
  }
}

foo := &Foo{
  Plain: "foo",
  Nested: &Bar{
    Plain: "bar",
  },
}

The result of DecodeBody will be:

foo := &Foo{
  Plain: "foo",
  Nested: &Bar{
    More: "foo",
  },
}

With this commit, decode functions will check will check if the value is
nil and only then set them to empty struct, which allows for appending
to existing structs.

In case of a slice, either new empty element will be added, or existing
element will be used for setting new value.

Result after this patch:

foo := &Foo{
  Plain: "foo",
  Nested: &Bar{
    More: "foo",
    Plain: "bar,
  },
}

While it is not very useful for slices, since default values would have
to be positional, it is useful for standalone blocks and it makes
decoding process to behave more consistent.

Alternative to this patch would be always zeroing top-level struct when
decode process starts, to maintain clear and consistent behaviour.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>