Add test cases for multiline heredocs
These tests pass, but they close what's effectively a gap in the test
coverage wherein nothing was ensuring that the whitespace prefix
trimming behavior still gave the same/correct output with blank lines
in the middle.

I made these tests because nomad templates were behaving oddly, only
to discover that nomad has a mix of both hcl1 and hcl2 in it!
Fortunately, hcl2's behavior is correct, but it should stay that way.